# doc/specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Store Daily QA System API",
    "description": "API for store staff daily QA system",
    "version": "1.0.0"
  },
  "paths": {
    "/api/admin/survey/questions": {
      "get": {
        "tags": ["admin"],
        "summary": "Get Questions",
        "description": "定型質問一覧取得",
        "operationId": "get_questions_api_admin_survey_questions_get",
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "アクティブな質問のみ取得",
              "default": true,
              "title": "Is Active"
            },
            "description": "アクティブな質問のみ取得"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "取得件数上限",
              "default": 30,
              "title": "Limit"
            },
            "description": "取得件数上限"
          }
        ],
        "responses": {
          "200": {
            "description": "Questions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      },
      "post": {
        "tags": ["admin"],
        "summary": "Create Question",
        "description": "定型質問作成",
        "operationId": "create_question_api_admin_survey_questions_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QuestionCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOperationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/api/admin/survey/questions/{admin_survey_id}": {
      "get": {
        "tags": ["admin"],
        "summary": "Get Question",
        "description": "定型質問詳細取得",
        "operationId": "get_question_api_admin_survey_questions__admin_survey_id__get",
        "parameters": [
          {
            "name": "admin_survey_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Admin Survey Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Question retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/QuestionResponse" }
              }
            }
          },
          "404": {
            "description": "Question not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      },
      "put": {
        "tags": ["admin"],
        "summary": "Update Question",
        "description": "定型質問更新",
        "operationId": "update_question_api_admin_survey_questions__admin_survey_id__put",
        "parameters": [
          {
            "name": "admin_survey_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Admin Survey Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QuestionUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOperationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "404": {
            "description": "Question not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      },
      "delete": {
        "tags": ["admin"],
        "summary": "Delete Question",
        "description": "定型質問削除",
        "operationId": "delete_question_api_admin_survey_questions__admin_survey_id__delete",
        "parameters": [
          {
            "name": "admin_survey_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Admin Survey Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Question deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOperationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Question not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/api/admin/messages": {
      "get": {
        "tags": ["admin"],
        "summary": "Get Messages",
        "description": "お知らせ一覧取得",
        "operationId": "get_messages_api_admin_messages_get",
        "parameters": [
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "アクティブなお知らせのみ取得",
              "default": true,
              "title": "Is Active"
            },
            "description": "アクティブなお知らせのみ取得"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "取得件数上限",
              "default": 30,
              "title": "Limit"
            },
            "description": "取得件数上限"
          }
        ],
        "responses": {
          "200": {
            "description": "Messages retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageListResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      },
      "post": {
        "tags": ["admin"],
        "summary": "Create Message",
        "description": "お知らせ作成",
        "operationId": "create_message_api_admin_messages_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "title": "Authorization" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MessageCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOperationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
    "/api/admin/messages/{admin_messages_id}": {
      "get": {
        "tags": ["admin"],
        "summary": "Get Message",
        "description": "お知らせ詳細取得",
        "operationId": "get_message_api_admin_messages__admin_messages_id__get",
        "parameters": [
          {
            "name": "admin_messages_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Admin Messages Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Message retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "404": {
            "description": "Message not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      },
      "put": {
        "tags": ["admin"],
        "summary": "Update Message",
        "description": "お知らせ更新",
        "operationId": "update_message_api_admin_messages__admin_messages_id__put",
        "parameters": [
          {
            "name": "admin_messages_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Admin Messages Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MessageUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOperationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "404": {
            "description": "Message not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      },
      "delete": {
        "tags": ["admin"],
        "summary": "Delete Message",
        "description": "お知らせ削除",
        "operationId": "delete_message_api_admin_messages__admin_messages_id__delete",
        "parameters": [
          {
            "name": "admin_messages_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Admin Messages Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Message deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOperationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Message not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/APIError" }
              }
            }
          },
          "422": { "description": "Validation error" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorResponse" }
        },
        "type": "object",
        "required": ["error"],
        "title": "APIError",
        "description": "API エラーレスポンス"
      },
      "AdminOperationResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作結果メッセージ"
          }
        },
        "type": "object",
        "required": ["message"],
        "title": "AdminOperationResponse",
        "description": "管理者操作レスポンス"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "エラーコード"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "エラーメッセージ"
          },
          "details": {
            "anyOf": [{}, { "type": "null" }],
            "title": "Details",
            "description": "エラー詳細"
          }
        },
        "type": "object",
        "required": ["code", "message"],
        "title": "ErrorResponse",
        "description": "エラーレスポンス"
      },
...
      "MessageCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "description": "タイトル"
          },
          "content": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content",
            "description": "内容"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "アクティブ状態",
            "default": true
          }
        },
        "type": "object",
        "required": ["title", "content"],
        "title": "MessageCreateRequest",
        "description": "お知らせ作成リクエスト"
      },
...
      "MessageListResponse": {
        "properties": {
          "messages": {
            "items": { "$ref": "#/components/schemas/MessageResponse" },
            "type": "array",
            "title": "Messages",
            "description": "お知らせリスト"
          }
        },
        "type": "object",
        "required": ["messages"],
        "title": "MessageListResponse",
        "description": "お知らせ一覧レスポンス"
      },
      "MessageOperationResponse": {
        "properties": {
          "admin_messages_id": {
            "type": "string",
            "title": "Admin Messages Id",
            "description": "作成/更新されたメッセージID"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作結果メッセージ"
          }
        },
        "type": "object",
        "required": ["admin_messages_id", "message"],
        "title": "MessageOperationResponse",
        "description": "メッセージ操作レスポンス"
      },
      "MessageResponse": {
        "properties": {
          "admin_messages_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Admin Messages Id",
            "description": "お知らせID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "タイトル"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "内容"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "アクティブ状態"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "作成日時"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新日時"
          }
        },
        "type": "object",
        "required": [
          "admin_messages_id",
          "title",
          "content",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "MessageResponse",
        "description": "お知らせレスポンス"
      },
      "MessageUpdateRequest": {
        "properties": {
          "title": {
            "anyOf": [
              { "type": "string", "maxLength": 200 },
              { "type": "null" }
            ],
            "title": "Title",
            "description": "タイトル"
          },
          "content": {
            "anyOf": [
              { "type": "string", "maxLength": 2000 },
              { "type": "null" }
            ],
            "title": "Content",
            "description": "内容"
          },
          "is_active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Active",
            "description": "アクティブ状態"
          }
        },
        "type": "object",
        "title": "MessageUpdateRequest",
        "description": "お知らせ更新リクエスト"
      },
      "QuestionCreateRequest": {
        "properties": {
          "question_text": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Question Text",
            "description": "質問内容"
          },
          "question_type": { "$ref": "#/components/schemas/QuestionType" },
          "options": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Options",
            "description": "選択肢(choiceタイプの場合)"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "アクティブ状態",
            "default": true
          }
        },
        "type": "object",
        "required": ["question_text", "question_type"],
        "title": "QuestionCreateRequest",
        "description": "定型質問作成リクエスト"
      },
      "QuestionListResponse": {
        "properties": {
          "questions": {
            "items": { "$ref": "#/components/schemas/QuestionResponse" },
            "type": "array",
            "title": "Questions",
            "description": "質問リスト"
          }
        },
        "type": "object",
        "required": ["questions"],
        "title": "QuestionListResponse",
        "description": "定型質問一覧レスポンス"
      },
      "QuestionOperationResponse": {
        "properties": {
          "admin_survey_id": {
            "type": "string",
            "title": "Admin Survey Id",
            "description": "作成/更新された質問ID"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作結果メッセージ"
          }
        },
        "type": "object",
        "required": ["admin_survey_id", "message"],
        "title": "QuestionOperationResponse",
        "description": "質問操作レスポンス"
      },
      "QuestionResponse": {
        "properties": {
          "admin_survey_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Admin Survey Id",
            "description": "質問ID"
          },
          "question_text": {
            "type": "string",
            "title": "Question Text",
            "description": "質問内容"
          },
          "question_type": { "$ref": "#/components/schemas/QuestionType" },
          "options": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Options",
            "description": "選択肢"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "アクティブ状態"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "作成日時"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新日時"
          }
        },
        "type": "object",
        "required": [
          "admin_survey_id",
          "question_text",
          "question_type",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "QuestionResponse",
        "description": "定型質問レスポンス"
      },
      "QuestionType": {
        "type": "string",
        "enum": ["text", "choice", "rating", "number"],
        "title": "QuestionType",
        "description": "質問タイプ"
      },
      "QuestionUpdateRequest": {
        "properties": {
          "question_text": {
            "anyOf": [
              { "type": "string", "maxLength": 500 },
              { "type": "null" }
            ],
            "title": "Question Text",
            "description": "質問内容"
          },
          "question_type": {
            "anyOf": [
              { "$ref": "#/components/schemas/QuestionType" },
              { "type": "null" }
            ],
            "description": "質問タイプ"
          },
          "options": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Options",
            "description": "選択肢(choiceタイプの場合)"
          },
          "is_active": {
            "anyOf": [{ "type": "boolean" }, { "type": "null" }],
            "title": "Is Active",
            "description": "アクティブ状態"
          }
        },
        "type": "object",
        "title": "QuestionUpdateRequest",
        "description": "定型質問更新リクエスト"
      },
...
    }
  }
}